Bump eslint in our 3 config packages (#62)
Bumping the following eslint-related pacakges used in our configurations:

-  `eslint`: `^3.1.1` -> `^4.19.1`
- `babel-eslint`: `^7.1.1` -> `^8.2.3`
- `eslint-plugin-import`: `^2.2.0` -> `^2.11.0`
- `eslint-plug-react`: `^6.7.1` -> `^7.7.0`
- `eslint-plugin-jsx-a11y`: `^6.0.0` -> `^6.0.3`

As a result, I was able to add a bunch more new rules in `eslint-config-eventbrite-legacy`, `eslint-config-eventbrite` & `eslint-config-eventbrite-react`. `eslint-plugin-jsx-a11y` has a "strict" configuration, so was able to get rid of a lot of rules in place of just extending it.

For `eslint-config-eventbrite`, I also added `eslint-plugin-jest@^21.15.1` so that we can lint against our Jest test as well.

In order to make this all easier to use, I converted the `packages/` folder over to [yarn workspaces](https://yarnpkg.com/lang/en/docs/workspaces/). This made it easier to maintain the linked dependencies. This should also makes it easier to run tests against them in Travis.

Lastly, in order to verify that all of the rules are valid, each configuration lints (`yarn run lint`) against the rules it defines.

Fixes #56.