Remove travis-deploy-once in favor of node filter
For some reason `travis-deploy-once` needs `GITHUB_TOKEN` set in the environment (or passed in) in order to work, which makes no sense because we won't be writing anything to Github. It's probably because it's typically used with `semantic-release` which _does_ write to Github.

So instead, going to use the `node` conditional deployment filter and have it only deploy for the `"node"` job (i.e. latest version of node).