Switch to npm publish instead of yarn publish
`yarn publish` for some reason asks for the new version of the package, makes it pretty much useless in a CI environment. I want it to just publish the files at the version _already_ in the `package.json`. This is what `npm publish` does so switching to that in hopes that it works as intended.