Add .npmrc file to set authToken
[`Build #88`](https://travis-ci.org/eventbrite/javascript/jobs/383401916) failed because of an authorization error.

In Travis we have `NPM_TOKEN` as an environment variable, but it's not automagically read. We need to update the registry URL to include it per: https://docs.npmjs.com/private-modules/ci-server-config#setting-up-a-project-specific-npmrc-file.

Then `npm publish` should have the authorization it needs. Hoping that having the one `.npmrc` file in the project root will be enough.