Add .npmrc file to set authToken (#70)
[`Build #88`](https://travis-ci.org/eventbrite/javascript/jobs/383401916) failed because of an authorization error.

In Travis we have `NPM_TOKEN` as an environment variable, but it's not automagically read. We need to update the registry URL to include it per: https://docs.npmjs.com/private-modules/ci-server-config#setting-up-a-project-specific-npmrc-file.

Then `npm publish` should have the authorization it needs.

Tried committing a `.npmrc` file (see [`Build #89`](https://travis-ci.org/eventbrite/javascript/builds/383421546)), but that broke PRs because `NPM_TOKEN` isn't available in builds from PRs the outside. So instead in `before_deploy` we'll create the `.npmrc` file on the fly. This is was `semantic-release` and other utilities do behind the scenes.