Add skip_cleanup to each deploy
Based on [`Build #92`](https://travis-ci.org/eventbrite/javascript/jobs/383442045), Travis cleans up the build artificats (`git stash --all`) **after** `before_deploy` is run and before `deploy` is executed. This means that the temporary `.npmrc` file that was written with #70 was removed before deploying. As a result no authentication again.

Passing `skip_cleanup: true` should do the trick. Not sure if it needs to be passed to each or if it could be shared at the top under `deploy`, but going the safe route and adding to each.